Extend support backwards to Node v14
For a project like this, we want to sensibly support as many users
as we can. Node v14 is still LTS, so we should claim support for
it (and it's easy for us to do so). Fastify v4 also tests against
Node v14. Next year, v14 will be EOL at which time we can drop
support for it in a major version release.

Since we support back to v14, we should downgrade the node types
to match so that we get typescript errors if we try to use > v14
Node APIs.

I also added "diagnostics: true" to the ts-jest config so we get
test failures when we have TS errors in test files.