Remove graphql as a peer dependency
The graphql peer dep is specified by Apollo Server and doesn't
need to be specified here as well. Since Apollo Server is a
peer dep of this package, graphql will transitively be required
in projects that use this package.

This means we don't duplicate the peer dependency requirement
and potentially fall out of sync with Apollo Server's range
of acceptable versions.